\name{glm.predict-package}
\alias{glm.predict-package}
\alias{glm.predict}
\docType{package}
\title{
Predicted Values and Discrete Changes for GLM
}
\description{
This package provides functions to calculate predicted values
    and the difference between two cases with confidence interval.
}

\author{
Benjamin Schlegel

Maintainer: Benjamin Schlegel <kontakt@benjaminschlegel.ch>
}

\keyword{ package }
